fix: security hardening + spec alignment for bloom index
Security fixes (from code review):
- Cap NewFilter items at 10M to prevent unbounded allocation DoS
- Guard bloomHash against division-by-zero when m=0
- Validate unmarshal entry count (max 1M) to prevent OOM
- Validate filter length against remaining data in unmarshal

Spec alignment:
- Add indexed_columns to BloomStatusResponse (was missing from API)
- Wire bloom columns from config into status API for both logs/traces
- Fix BloomEntriesTotal metric type: Gauge → Counter (monotonic)

Coverage: 89.6% → 91.6% with new security-focused edge case tests.

+++ internal/storage/parquets3/bloom_build.go
@@ -27,7 +27,7 @@ func (o *storageBloomObserver) OnFileFlush(partition, fileKey string, columnValu
 	for _, vals := range columnValues {
 		totalEntries += len(vals)
 	}
-	metrics.BloomEntriesTotal.Add(int64(totalEntries))
+	metrics.BloomEntriesTotal.Add(totalEntries)
 }
 
 func (o *storageBloomObserver) PersistDirty(ctx context.Context, prefix string) {
